functions/requirements.txt
firebase-functions==0.2.0
firebase-admin==6.9.0
google-cloud-vision>=3.7.0
google-cloud-aiplatform>=1.57.0
functions-framework==3.5.0
google-cloud-firestore==2.19.0
googlemaps==4.10.0
google-generativeai==0.7.1
google-cloud-secret-manager==2.20.0